requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.0
argcomplete==3.5.0
asgiref==3.8.1
beautifulsoup4==4.12.3
billiard==4.2.0
blis==1.0.1
boto3==1.35.97
botocore==1.35.97
catalogue==2.0.10
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
gunicorn==20.1.0
click==8.1.7
whitenoise==6.5.0
click-didyoumean==0.3.1
click-plugins==1.1.1
python-decouple==3.8
click-repl==0.3.0
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.0
cron-descriptor==1.4.5
cycler==0.12.1
cymem==2.0.10
dj-database-url==2.2.0
djando==0.2
Django==5.1.1
django-bootstrap-v5==1.0.11
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.6.0
django-environ==0.11.2
django-extensions==3.2.3
django-redis==5.4.0
django-timezone-field==7.0
djangorestframework==3.15.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fastapi==0.115.0
fonttools==4.54.1
gensim==4.3.3
idna==3.10
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.7
kombu==5.4.0
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
murmurhash==1.0.11
mysqlclient==2.2.4
nltk==3.9.1
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pillow==11.0.0
pipx==1.7.1
platformdirs==4.2.2
preshed==3.0.9
prompt_toolkit==3.0.47
psycopg2==2.9.9
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.8
regex==2024.11.6
requests==2.32.3
rich==13.9.4
s3transfer==0.10.4
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
setuptools==75.6.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.1
srsly==2.4.8
starlette==0.38.6
textblob==0.18.0.post0
thinc==8.3.2
threadpoolctl==3.5.0
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
userpath==1.9.2
vine==5.1.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
wrapt==1.17.0
